Rename nconvert.py -> convert.py
