+++ python/src/scippneutron/convert.py
